requirements.txt
# Core Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# PDF Generation
weasyprint>=60.0
pillow>=10.1.0

# Data Processing
pandas>=2.1.0
openpyxl>=3.1.0  # For Excel support
python-multipart>=0.0.6

# AI/LLM
google-generativeai>=0.3.0
PyJWT>=2.8.0
email-validator>=2.1.0  # Recommended: Google Gemini API
openai>=1.3.0  # Optional: OpenAI GPT-4
anthropic>=0.7.0  # Optional: Claude API
langchain>=0.1.0  # Optional: for advanced LLM workflows

# Task Queue (Optional for Phase 5)
celery>=5.3.0
redis>=5.0.0

# Database (Optional for Phase 5)
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Font Handling
fonttools>=4.47.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
click>=8.1.7  # For CLI
rich>=13.7.0  # For beautiful CLI output
tqdm>=4.66.0  # Progress bars

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0  # For API testing

# Development
black>=23.11.0
ruff>=0.1.6
mypy>=1.7.0

# Optional: Web UI (Streamlit)
streamlit>=1.28.0

# Optional: Image Processing
requests>=2.31.0  # For downloading images
Pillow>=10.1.0  # For image manipulation
beautifulsoup4>=4.12.0  # For web scraping HTML parsing

